[feat] : 보틀 보관함 UI 구현

+++ feat/bottle/src/main/kotlin/com/team/bottles/feat/bottle/navigation/BottleNavigation.kt
@@ -4,6 +4,7 @@ import BottleNavigator
 import MainNavigator
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
+import androidx.navigation.toRoute
 import com.team.bottles.feat.bottle.BottleRoute
 import com.team.bottles.feat.bottle.arrivedbottles.ArrivedBottlesRoute
 import com.team.bottles.feat.bottle.bottlebox.BottleBoxRoute
@@ -17,7 +18,7 @@ fun NavGraphBuilder.arrivedBottlesScreen(
 }
 
 fun NavGraphBuilder.bottleBoxScreen(
-    navigateToBottle: () -> Unit
+    navigateToBottle: (Int) -> Unit
 ) {
     composable<MainNavigator.BottlesBox> {
         BottleBoxRoute(navigateToBottle = navigateToBottle)
